reuse free blocks

+++ malloc.h
@@ -8,7 +8,7 @@
 # ifndef PAGE_SIZE
 #  define PAGE_SIZE			((unsigned)getpagesize())
 # endif
-# define ALIGN_SIZE			__alignof__(unsigned long)
+# define ALIGN_SIZE			__alignof__(unsigned long long)
 # define IS_DISALIGNED(x)	((x) % ALIGN_SIZE)
 # define DO_ALIGN(x)		((x) + ALIGN_SIZE - ((x) % ALIGN_SIZE))
 # define ALIGN(x)			(IS_DISALIGNED(x) ? DO_ALIGN(x) : (x))
